Main Page, Reference to JFrame
Changed admin frame to include a jframe only, with the jpanel on a seperate class called "Main Page".
also had an issue with the mainpage panel while opening the AddBusiness jpanel,
fixed it by adding a reference to the admin jframe to each and everyone of the jpanels that is sent by the constructor.